requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
comet-ml
dvc
dvc-gs
flask
joblib
kagglehub
scikit-learn
matplotlib
utils
python-dotenv
boto3
botocore